Fixed Language dropdown and footer + QoL change
The language dropdown no longer falls outside the navigation bar.

The page footer no longer hides page contents.
Replaced hardcoded <br>s and <div>s by CSS values in doing so.

Added a background colour to the body that is similar to the background image. This mitigates the former white-flash when the background image was not yet loaded upon opening the page.

Removed some unused CSS.